+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- CLS-03: schedule/expiry on drafts; acknowledgement counts for teachers
 - CLS-02: teachers attach https links to draft classroom announcements
 - CLS-01: teachers create draft classroom announcements for assigned scopes
 - MRK-05: teachers see class performance summaries for published assessments